Added support for HMServiceGroups
