+++ LightingKit/Strategies/FilterStrategy.swift
