+++ LightingKit/Strategies/SuccessStrategy.swift
