+++ LightingKit/Swift Extensions/FoundationExtensions.swift
