+++ LightingKit/Swift Extensions/Int+SanitizeBrightness.swift
