+++ LightingKit/Swift Extensions/TimeInterval+FadeBrightness.swift
